@@ -3,9 +3,9 @@
  */
 
 import * as AWS from "aws-sdk"; // imported for Types
-import DomainInfo = require("./DomainInfo");
-import Globals from "./Globals";
-import {CustomDomain} from "./types";
+import DomainInfo = require("./domain-info");
+import Globals from "./globals";
+import {CustomDomain, Route53Params} from "./types";
 
 class DomainConfig {
 
@@ -27,11 +27,13 @@ class DomainConfig {
     public securityPolicy: string | undefined;
     public autoDomain: boolean | undefined;
     public autoDomainWaitFor: string | undefined;
+    public route53Params: Route53Params;
 
     public domainInfo: DomainInfo | undefined;
     public apiId: string | undefined;
     public apiMapping: AWS.ApiGatewayV2.GetApiMappingResponse;
     public allowPathMatching: boolean | false;
+    public region: string;
 
     constructor(config: CustomDomain) {
 
@@ -82,12 +84,35 @@ class DomainConfig {
         }
         this.securityPolicy = tlsVersionToUse;
 
-        let region = Globals.defaultRegion;
+        this.region = Globals.defaultRegion;
         if (this.endpointType === Globals.endpointTypes.regional) {
-            region = Globals.serverless.providers.aws.getRegion();
+            this.region = Globals.serverless.providers.aws.getRegion();
         }
-        const acmCredentials = Object.assign({}, Globals.serverless.providers.aws.getCredentials(), {region});
+        const acmCredentials = Object.assign(
+            {}, Globals.serverless.providers.aws.getCredentials(), {region: this.region}
+        );
         this.acm = new Globals.serverless.providers.aws.sdk.ACM(acmCredentials);
+
+        const defaultRoutingPolicy = Globals.routingPolicies.simple;
+        const routingPolicy = config.route53Params?.routingPolicy?.toLowerCase() ?? defaultRoutingPolicy;
+        const routingPolicyToUse = Globals.routingPolicies[routingPolicy];
+        if (!routingPolicyToUse) {
+            throw new Error(`${routingPolicy} is not a supported routing policy, use simple, latency, or weighted.`);
+        }
+
+        if (routingPolicyToUse !== defaultRoutingPolicy && endpointTypeToUse === Globals.endpointTypes.edge) {
+            throw new Error(
+                `${routingPolicy} routing is not intended to be used with edge endpoints. ` +
+                "Use a regional endpoint instead."
+            );
+        }
+
+        this.route53Params = {
+            routingPolicy: routingPolicyToUse,
+            setIdentifier: config.route53Params?.setIdentifier,
+            weight: config.route53Params?.weight ?? 200,
+            healthCheckId: config.route53Params?.healthCheckId
+        }
     }
 
     /**
@@ -100,7 +125,6 @@ class DomainConfig {
      * If no customDomain object exists, an exception is thrown.
      */
     private evaluateEnabled(enabled: any): boolean {
-        // const enabled = this.serverless.service.custom.customDomain.enabled;
         if (enabled === undefined) {
             return true;
         }
