@@ -7,7 +7,7 @@ module.exports.helloWorld = (event, context, callback) => {
       "Access-Control-Allow-Origin": "*", // Required for CORS support to work
     },
     body: JSON.stringify({
-      message: "Go Serverless v1.0! Your function executed successfully!",
+      message: "Go Serverless! Your function executed successfully!",
       input: event,
     }),
   };
